+++ test/stress_tests.php
@@ -252,6 +252,66 @@
                         echo('<h5>The test took '. sprintf('%01.3f', microtime(TRUE) - $start) . ' seconds.</h5>');
                     echo('</div>');
                 echo('</div>');
+                echo('<div id="test-037" class="inner_closed">');
+                    echo('<h2 class="inner_header"><a href="javascript:toggle_inner_state(\'test-037\')">TEST 37: Look for multiple meeting names, logged in as BillyBob.</a></h2>');
+
+                    echo('<div class="main_div inner_container">');
+                        ?>
+                        <div class="main_div" style="margin-right:2em">
+                        <p class="explain">This tests looks for meetings named "Just for Today", "Back to Basics" or "Free at Last".</p>
+                        <p class="explain">We expect this test to succeed.</p>
+                        </div>
+                        <?php
+                        $start = microtime(TRUE);
+                        try_custom_query_5('billybob', '', 'CoreysGoryStory');
+                        echo('<h5>The test took '. sprintf('%01.3f', microtime(TRUE) - $start) . ' seconds.</h5>');
+                    echo('</div>');
+                echo('</div>');
+                echo('<div id="test-038" class="inner_closed">');
+                    echo('<h2 class="inner_header"><a href="javascript:toggle_inner_state(\'test-038\')">TEST 38: Look for multiple states in tags, logged in as BillyBob.</a></h2>');
+
+                    echo('<div class="main_div inner_container">');
+                        ?>
+                        <div class="main_div" style="margin-right:2em">
+                        <p class="explain">This tests looks for meetings with state tags of "HI", "KS" or "TN".</p>
+                        <p class="explain">We expect this test to succeed.</p>
+                        </div>
+                        <?php
+                        $start = microtime(TRUE);
+                        try_custom_query_6('billybob', '', 'CoreysGoryStory');
+                        echo('<h5>The test took '. sprintf('%01.3f', microtime(TRUE) - $start) . ' seconds.</h5>');
+                    echo('</div>');
+                echo('</div>');
+                echo('<div id="test-039" class="inner_closed">');
+                    echo('<h2 class="inner_header"><a href="javascript:toggle_inner_state(\'test-039\')">TEST 39: Look for multiple contiguous IDs, logged in as BillyBob.</a></h2>');
+
+                    echo('<div class="main_div inner_container">');
+                        ?>
+                        <div class="main_div" style="margin-right:2em">
+                        <p class="explain">This tests looks for meetings with IDs from 1 - 31.</p>
+                        <p class="explain">We expect this test to succeed, but we should never get ID "1".</p>
+                        </div>
+                        <?php
+                        $start = microtime(TRUE);
+                        try_custom_query_7('billybob', '', 'CoreysGoryStory');
+                        echo('<h5>The test took '. sprintf('%01.3f', microtime(TRUE) - $start) . ' seconds.</h5>');
+                    echo('</div>');
+                echo('</div>');
+                echo('<div id="test-040" class="inner_closed">');
+                    echo('<h2 class="inner_header"><a href="javascript:toggle_inner_state(\'test-040\')">TEST 40: Look for multiple random IDs, logged in as BillyBob.</a></h2>');
+
+                    echo('<div class="main_div inner_container">');
+                        ?>
+                        <div class="main_div" style="margin-right:2em">
+                        <p class="explain">This tests looks for meetings with many different IDs (including ones out of range).</p>
+                        <p class="explain">We expect this test to succeed, but we should never get ID "1".</p>
+                        </div>
+                        <?php
+                        $start = microtime(TRUE);
+                        try_custom_query_8('billybob', '', 'CoreysGoryStory');
+                        echo('<h5>The test took '. sprintf('%01.3f', microtime(TRUE) - $start) . ' seconds.</h5>');
+                    echo('</div>');
+                echo('</div>');
             echo('</div>');
         echo('</div>');
     ?>
@@ -377,6 +437,146 @@
             echo("<h2>The access instance is valid!</h2>");
             $st1 = microtime(TRUE);
             $test_item = $access_instance->generic_search(Array('location' => Array('longitude' => -95.3698, 'latitude' => 29.7604, 'radius' => 100.0)));
+            $fetchTime = sprintf('%01.3f', microtime(TRUE) - $st1);
+            echo('<div class="inner_div">');
+                if ( isset($test_item) ) {
+                    if (is_array($test_item)) {
+                        if (count($test_item)) {
+                            echo("<h4>We got ".count($test_item)." records in $fetchTime seconds.</h4>");
+                            $count = 0;
+                            foreach ( $test_item as $item ) {
+                                display_record($item);
+                            }
+                        }
+                    }
+                }
+            echo('</div>');
+        } else {
+            echo("<h2 style=\"color:red;font-weight:bold\">The access instance is not valid!</h2>");
+            echo('<p style="margin-left:1em;color:red;font-weight:bold">Error: ('.$access_instance->error->error_code.') '.$access_instance->error->error_name.' ('.$access_instance->error->error_description.')</p>');
+        }
+    }
+    
+    function try_custom_query_5($in_login = NULL, $in_hashed_password = NULL, $in_password = NULL) {
+        $access_instance = NULL;
+        
+        if ( !defined('LGV_ACCESS_CATCHER') ) {
+            define('LGV_ACCESS_CATCHER', 1);
+        }
+        
+        require_once(CO_Config::main_class_dir().'/co_access.class.php');
+        
+        $access_instance = new CO_Access($in_login, $in_hashed_password, $in_password);
+        
+        if ($access_instance->valid) {
+            echo("<h2>The access instance is valid!</h2>");
+            $st1 = microtime(TRUE);
+            $test_item = $access_instance->generic_search(Array('name' => Array('back to basics', 'just for today', 'free at last')));
+            $fetchTime = sprintf('%01.3f', microtime(TRUE) - $st1);
+            echo('<div class="inner_div">');
+                if ( isset($test_item) ) {
+                    if (is_array($test_item)) {
+                        if (count($test_item)) {
+                            echo("<h4>We got ".count($test_item)." records in $fetchTime seconds.</h4>");
+                            $count = 0;
+                            foreach ( $test_item as $item ) {
+                                display_record($item);
+                            }
+                        }
+                    }
+                }
+            echo('</div>');
+        } else {
+            echo("<h2 style=\"color:red;font-weight:bold\">The access instance is not valid!</h2>");
+            echo('<p style="margin-left:1em;color:red;font-weight:bold">Error: ('.$access_instance->error->error_code.') '.$access_instance->error->error_name.' ('.$access_instance->error->error_description.')</p>');
+        }
+    }
+    
+    function try_custom_query_6($in_login = NULL, $in_hashed_password = NULL, $in_password = NULL) {
+        $access_instance = NULL;
+        
+        if ( !defined('LGV_ACCESS_CATCHER') ) {
+            define('LGV_ACCESS_CATCHER', 1);
+        }
+        
+        require_once(CO_Config::main_class_dir().'/co_access.class.php');
+        
+        $access_instance = new CO_Access($in_login, $in_hashed_password, $in_password);
+        
+        if ($access_instance->valid) {
+            echo("<h2>The access instance is valid!</h2>");
+            $st1 = microtime(TRUE);
+            $test_item = $access_instance->generic_search(Array('tags' => Array('', Array('HI','KS','TN'))));
+            $fetchTime = sprintf('%01.3f', microtime(TRUE) - $st1);
+            echo('<div class="inner_div">');
+                if ( isset($test_item) ) {
+                    if (is_array($test_item)) {
+                        if (count($test_item)) {
+                            echo("<h4>We got ".count($test_item)." records in $fetchTime seconds.</h4>");
+                            $count = 0;
+                            foreach ( $test_item as $item ) {
+                                display_record($item);
+                            }
+                        }
+                    }
+                }
+            echo('</div>');
+        } else {
+            echo("<h2 style=\"color:red;font-weight:bold\">The access instance is not valid!</h2>");
+            echo('<p style="margin-left:1em;color:red;font-weight:bold">Error: ('.$access_instance->error->error_code.') '.$access_instance->error->error_name.' ('.$access_instance->error->error_description.')</p>');
+        }
+    }
+    
+    function try_custom_query_7($in_login = NULL, $in_hashed_password = NULL, $in_password = NULL) {
+        $access_instance = NULL;
+        
+        if ( !defined('LGV_ACCESS_CATCHER') ) {
+            define('LGV_ACCESS_CATCHER', 1);
+        }
+        
+        require_once(CO_Config::main_class_dir().'/co_access.class.php');
+        
+        $access_instance = new CO_Access($in_login, $in_hashed_password, $in_password);
+        
+        if ($access_instance->valid) {
+            echo("<h2>The access instance is valid!</h2>");
+            $st1 = microtime(TRUE);
+            $test_item = $access_instance->generic_search(Array('id' => Array(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31)));
+            $fetchTime = sprintf('%01.3f', microtime(TRUE) - $st1);
+            echo('<div class="inner_div">');
+                if ( isset($test_item) ) {
+                    if (is_array($test_item)) {
+                        if (count($test_item)) {
+                            echo("<h4>We got ".count($test_item)." records in $fetchTime seconds.</h4>");
+                            $count = 0;
+                            foreach ( $test_item as $item ) {
+                                display_record($item);
+                            }
+                        }
+                    }
+                }
+            echo('</div>');
+        } else {
+            echo("<h2 style=\"color:red;font-weight:bold\">The access instance is not valid!</h2>");
+            echo('<p style="margin-left:1em;color:red;font-weight:bold">Error: ('.$access_instance->error->error_code.') '.$access_instance->error->error_name.' ('.$access_instance->error->error_description.')</p>');
+        }
+    }
+    
+    function try_custom_query_8($in_login = NULL, $in_hashed_password = NULL, $in_password = NULL) {
+        $access_instance = NULL;
+        
+        if ( !defined('LGV_ACCESS_CATCHER') ) {
+            define('LGV_ACCESS_CATCHER', 1);
+        }
+        
+        require_once(CO_Config::main_class_dir().'/co_access.class.php');
+        
+        $access_instance = new CO_Access($in_login, $in_hashed_password, $in_password);
+        
+        if ($access_instance->valid) {
+            echo("<h2>The access instance is valid!</h2>");
+            $st1 = microtime(TRUE);
+            $test_item = $access_instance->generic_search(Array('id' => Array(1,2056,32,4789,599,6542,7,8,9,10000,11432,121,13,4,156,16,17,18,19,20210,21,22,2376,2,254,26,26,26,276,28,229,300000,341)));
             $fetchTime = sprintf('%01.3f', microtime(TRUE) - $st1);
             echo('<div class="inner_div">');
                 if ( isset($test_item) ) {