+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+**1.0.0.1001** *(March 27, 2018)*
+
+- Added some tests for multiple values of generic search terms.
+- Fixed a few issues with multiple values of generic search terms.
+
 **1.0.0.1000** *(March 27, 2018)*
 
 - Initial Development Tag.